cql3: Remove term
term isn't used anywhere now. We can remove it and all classes that derive from it.
