+++ CMakeLists.txt
@@ -428,7 +428,6 @@ set(scylla_sources
     cql3/statements/update_statement.cc
     cql3/statements/use_statement.cc
     cql3/token_relation.cc
-    cql3/tuples.cc
     cql3/type_json.cc
     cql3/untyped_result_set.cc
     cql3/update_parameters.cc
