+++ configure.py
@@ -756,7 +756,6 @@ scylla_core = (['database.cc',
                 'cql3/keyspace_element_name.cc',
                 'cql3/lists.cc',
                 'cql3/sets.cc',
-                'cql3/tuples.cc',
                 'cql3/maps.cc',
                 'cql3/values.cc',
                 'cql3/expr/expression.cc',
