+++ cql3/functions/functions.cc
@@ -20,8 +20,6 @@
  */
 
 #include "functions.hh"
-
-#include "function_call.hh"
 #include "token_fct.hh"
 #include "cql3/maps.hh"
 #include "cql3/sets.hh"
@@ -65,9 +63,6 @@ bool abstract_function::requires_thread() const { return false; }
 bool as_json_function::requires_thread() const { return false; }
 
 thread_local std::unordered_multimap<function_name, shared_ptr<function>> functions::_declared = init();
-
-static bytes_opt execute_internal(cql_serialization_format sf, scalar_function& fun, std::vector<bytes_opt> params);
-static shared_ptr<terminal> make_terminal(shared_ptr<function> fun, cql3::raw_value result, cql_serialization_format sf);
 
 void functions::clear_functions() noexcept {
     functions::_declared = init();
@@ -444,217 +439,6 @@ functions::match_arguments(database& db, const sstring& keyspace,
 bool
 functions::type_equals(const std::vector<data_type>& t1, const std::vector<data_type>& t2) {
     return t1 == t2;
-}
-
-void
-function_call::fill_prepare_context(prepare_context& ctx) const {
-    if (ctx.is_processing_pk_restrictions() && !_fun->is_pure()) {
-        // Hacking around `const` specifier in the `collect_prepare_metadata`
-        // declaration since we also need to modify the current instance along
-        // with prepare metadata.
-        ctx.add_pk_function_call(static_pointer_cast<function_call>(
-            const_cast<function_call*>(this)->shared_from_this()));
-    }
-    for (auto&& t : _terms) {
-        t->fill_prepare_context(ctx);
-    }
-}
-
-shared_ptr<terminal>
-function_call::bind(const query_options& options) {
-    return make_terminal(_fun, cql3::raw_value::make_value(bind_and_get_internal(options)), options.get_cql_serialization_format());
-}
-
-cql3::raw_value_view
-function_call::bind_and_get_internal(const query_options& options) {
-    std::vector<bytes_opt> buffers;
-    buffers.reserve(_terms.size());
-    for (auto&& t : _terms) {
-        // For now, we don't allow nulls as argument as no existing function needs it and it
-        // simplify things.
-        auto val = expr::evaluate_to_raw_view(t, options);
-        if (!val) {
-            throw exceptions::invalid_request_exception(format("Invalid null value for argument to {}", *_fun));
-        }
-        buffers.push_back(to_bytes_opt(val));
-    }
-    if (_id) {
-        // Populate the cache only for LWT statements. Note that this code
-        // works only in places where `function_call::raw` AST nodes are
-        // created.
-        // These cases do not include selection clause in SELECT statement,
-        // hence no database inputs are possibly allowed to the functions
-        // evaluated here.
-        // We can cache every non-deterministic call here as this code branch
-        // acts the same way as if all arguments are equivalent to literal
-        // values at this point (already calculated).
-        auto val = options.find_cached_pk_function_call(*_id);
-        if (val) {
-            return raw_value_view::make_temporary(raw_value::make_value(*val));
-        }
-    }
-    auto result = execute_internal(options.get_cql_serialization_format(), *_fun, std::move(buffers));
-    if (_id) {
-        options.cache_pk_function_call(*_id, result);
-    }
-    return cql3::raw_value_view::make_temporary(cql3::raw_value::make_value(result));
-}
-
-expr::expression function_call::to_expression() {
-    std::vector<expr::expression> args;
-    args.reserve(_terms.size());
-
-    for (const ::shared_ptr<term>& t : _terms) {
-        args.emplace_back(expr::to_expression(t));
-    }
-
-    return expr::function_call {
-        .func = _fun,
-        .args = std::move(args),
-        .lwt_cache_id = ::make_shared<std::optional<uint8_t>>(_id),
-    };
-}
-
-static
-bytes_opt
-execute_internal(cql_serialization_format sf, scalar_function& fun, std::vector<bytes_opt> params) {
-    bytes_opt result = fun.execute(sf, params);
-    try {
-        // Check the method didn't lied on it's declared return type
-        if (result) {
-            fun.return_type()->validate(*result, sf);
-        }
-        return result;
-    } catch (marshal_exception& e) {
-        throw runtime_exception(format("Return of function {} ({}) is not a valid value for its declared return type {}",
-                                       fun, to_hex(result),
-                                       fun.return_type()->as_cql3_type()
-                                       ));
-    }
-}
-
-bool
-function_call::contains_bind_marker() const {
-    for (auto&& t : _terms) {
-        if (t->contains_bind_marker()) {
-            return true;
-        }
-    }
-    return false;
-}
-
-static
-shared_ptr<terminal>
-make_terminal(shared_ptr<function> fun, cql3::raw_value result, cql_serialization_format sf)  {
-    return visit(*fun->return_type(), make_visitor(
-    [&] (const list_type_impl& ltype) -> shared_ptr<terminal> {
-        return make_shared<lists::value>(lists::value::from_serialized(result.to_view(), ltype, sf));
-    },
-    [&] (const set_type_impl& stype) -> shared_ptr<terminal> {
-        return make_shared<sets::value>(sets::value::from_serialized(result.to_view(), stype, sf));
-    },
-    [&] (const map_type_impl& mtype) -> shared_ptr<terminal> {
-        return make_shared<maps::value>(maps::value::from_serialized(result.to_view(), mtype, sf));
-    },
-    [&] (const user_type_impl& utype) -> shared_ptr<terminal> {
-        // TODO (kbraun): write a test for this case when User Defined Functions are implemented
-        return make_shared<user_types::value>(user_types::value::from_serialized(result.to_view(), utype));
-    },
-    [&] (const abstract_type& type) -> shared_ptr<terminal> {
-        if (type.is_collection()) {
-            throw std::runtime_error(format("function_call::make_terminal: unhandled collection type {}", type.name()));
-        }
-        return make_shared<constants::value>(std::move(result), fun->return_type());
-    }
-    ));
-}
-
-expr::expression
-prepare_function_call(const expr::function_call& fc, database& db, const sstring& keyspace, lw_shared_ptr<column_specification> receiver) {
-    auto&& fun = std::visit(overloaded_functor{
-        [] (const shared_ptr<function>& func) {
-            return func;
-        },
-        [&] (const function_name& name) {
-            auto args = boost::copy_range<std::vector<::shared_ptr<assignment_testable>>>(fc.args | boost::adaptors::transformed(expr::as_assignment_testable));
-            auto fun = functions::functions::get(db, keyspace, name, args, receiver->ks_name, receiver->cf_name, receiver.get());
-            if (!fun) {
-                throw exceptions::invalid_request_exception(format("Unknown function {} called", name));
-            }
-            return fun;
-        },
-    }, fc.func);
-    if (fun->is_aggregate()) {
-        throw exceptions::invalid_request_exception("Aggregation function are not supported in the where clause");
-    }
-
-    // Can't use static_pointer_cast<> because function is a virtual base class of scalar_function
-    auto&& scalar_fun = dynamic_pointer_cast<scalar_function>(fun);
-
-    // Functions.get() will complain if no function "name" type check with the provided arguments.
-    // We still have to validate that the return type matches however
-    if (!receiver->type->is_value_compatible_with(*scalar_fun->return_type())) {
-        throw exceptions::invalid_request_exception(format("Type error: cannot assign result of function {} (type {}) to {} (type {})",
-                                                    fun->name(), fun->return_type()->as_cql3_type(),
-                                                    receiver->name, receiver->type->as_cql3_type()));
-    }
-
-    if (scalar_fun->arg_types().size() != fc.args.size()) {
-        throw exceptions::invalid_request_exception(format("Incorrect number of arguments specified for function {} (expected {:d}, found {:d})",
-                                                    fun->name(), fun->arg_types().size(), fc.args.size()));
-    }
-
-    std::vector<expr::expression> parameters;
-    parameters.reserve(fc.args.size());
-    bool all_terminal = true;
-    for (size_t i = 0; i < fc.args.size(); ++i) {
-        expr::expression e = prepare_expression(fc.args[i], db, keyspace, functions::make_arg_spec(receiver->ks_name, receiver->cf_name, *scalar_fun, i));
-        if (!expr::is<expr::constant>(e)) {
-            all_terminal = false;
-        }
-        parameters.push_back(std::move(e));
-    }
-
-    // If all parameters are terminal and the function is pure, we can
-    // evaluate it now, otherwise we'd have to wait execution time
-    expr::function_call fun_call {
-        .func = fun,
-        .args = std::move(parameters),
-        .lwt_cache_id = fc.lwt_cache_id
-    };
-    if (all_terminal && scalar_fun->is_pure()) {
-        return expr::evaluate(fun_call, query_options::DEFAULT);
-    } else {
-        return fun_call;
-    }
-}
-
-assignment_testable::test_result
-test_assignment_function_call(const cql3::expr::function_call& fc, database& db, const sstring& keyspace, const column_specification& receiver) {
-    // Note: Functions.get() will return null if the function doesn't exist, or throw is no function matching
-    // the arguments can be found. We may get one of those if an undefined/wrong function is used as argument
-    // of another, existing, function. In that case, we return true here because we'll throw a proper exception
-    // later with a more helpful error message that if we were to return false here.
-    try {
-        auto&& fun = std::visit(overloaded_functor{
-            [&] (const function_name& name) {
-                auto args = boost::copy_range<std::vector<::shared_ptr<assignment_testable>>>(fc.args | boost::adaptors::transformed(expr::as_assignment_testable));
-                return functions::get(db, keyspace, name, args, receiver.ks_name, receiver.cf_name, &receiver);
-            },
-            [] (const shared_ptr<function>& func) {
-                return func;
-            },
-        }, fc.func);
-        if (fun && receiver.type == fun->return_type()) {
-            return assignment_testable::test_result::EXACT_MATCH;
-        } else if (!fun || receiver.type->is_value_compatible_with(*fun->return_type())) {
-            return assignment_testable::test_result::WEAKLY_ASSIGNABLE;
-        } else {
-            return assignment_testable::test_result::NOT_ASSIGNABLE;
-        }
-    } catch (exceptions::invalid_request_exception& e) {
-        return assignment_testable::test_result::WEAKLY_ASSIGNABLE;
-    }
 }
 
 }
