+++ cql3/lists.cc
@@ -42,164 +42,6 @@ lists::uuid_index_spec_of(const column_specification& column) {
             ::make_shared<column_identifier>(format("uuid_idx({})", *column.name), true), uuid_type);
 }
 
-
-lists::value
-lists::value::from_serialized(const raw_value_view& val, const list_type_impl& type, cql_serialization_format sf) {
-    try {
-        utils::chunked_vector<managed_bytes_opt> elements;
-        if (sf.collection_format_unchanged()) {
-            utils::chunked_vector<managed_bytes> tmp = val.with_value([sf] (const FragmentedView auto& v) {
-                return partially_deserialize_listlike(v, sf);
-            });
-            elements.reserve(tmp.size());
-            for (auto&& element : tmp) {
-                elements.emplace_back(std::move(element));
-            }
-        } else [[unlikely]] {
-            auto l = val.deserialize<list_type_impl::native_type>(type, sf);
-            elements.reserve(l.size());
-            for (auto&& element : l) {
-                // elements can be null in lists that represent a set of IN values
-                elements.push_back(element.is_null() ? managed_bytes_opt() : managed_bytes_opt(type.get_elements_type()->decompose(element)));
-            }
-        }
-        return value(std::move(elements), type.shared_from_this());
-    } catch (marshal_exception& e) {
-        throw exceptions::invalid_request_exception(e.what());
-    }
-}
-
-cql3::raw_value
-lists::value::get(const query_options& options) {
-    return cql3::raw_value::make_value(get_with_protocol_version(cql_serialization_format::internal()));
-}
-
-managed_bytes
-lists::value::get_with_protocol_version(cql_serialization_format sf) {
-    // Can't use boost::indirect_iterator, because optional is not an iterator
-    auto deref = [] (managed_bytes_opt& x) { return *x; };
-    return collection_type_impl::pack_fragmented(
-            boost::make_transform_iterator(_elements.begin(), deref),
-            boost::make_transform_iterator( _elements.end(), deref),
-            _elements.size(), sf);
-}
-
-bool
-lists::value::equals(const list_type_impl& lt, const value& v) {
-    if (_elements.size() != v._elements.size()) {
-        return false;
-    }
-    return std::equal(_elements.begin(), _elements.end(),
-            v._elements.begin(),
-            [t = lt.get_elements_type()] (const managed_bytes_opt& e1, const managed_bytes_opt& e2) { return t->equal(*e1, *e2); });
-}
-
-sstring
-lists::value::to_string() const {
-    std::ostringstream os;
-    os << "[";
-    bool is_first = true;
-    for (auto&& e : _elements) {
-        if (!is_first) {
-            os << ", ";
-        }
-        is_first = false;
-        os << to_hex(e);
-    }
-    os << "]";
-    return os.str();
-}
-
-bool
-lists::delayed_value::contains_bind_marker() const {
-    // False since we don't support them in collection
-    return false;
-}
-
-void
-lists::delayed_value::fill_prepare_context(prepare_context& ctx) const {
-}
-
-shared_ptr<terminal>
-lists::delayed_value::bind(const query_options& options) {
-    utils::chunked_vector<managed_bytes_opt> buffers;
-    buffers.reserve(_elements.size());
-    for (auto&& t : _elements) {
-        auto bo = expr::evaluate_to_raw_view(t, options);
-
-        if (bo.is_null()) {
-            throw exceptions::invalid_request_exception("null is not supported inside collections");
-        }
-        if (bo.is_unset_value()) {
-            return constants::UNSET_VALUE;
-        }
-
-        buffers.push_back(bo.with_value([] (const FragmentedView auto& v) { return managed_bytes(v); }));
-    }
-    return ::make_shared<value>(buffers, _my_type);
-}
-
-shared_ptr<terminal>
-lists::delayed_value::bind_ignore_null(const query_options& options) {
-    utils::chunked_vector<managed_bytes_opt> buffers;
-    buffers.reserve(_elements.size());
-    for (auto&& t : _elements) {
-        auto bo = expr::evaluate_to_raw_view(t, options);
-
-        if (bo.is_null()) {
-            continue;
-        }
-        if (bo.is_unset_value()) {
-            return constants::UNSET_VALUE;
-        }
-
-        buffers.push_back(bo.with_value([] (const FragmentedView auto& v) { return managed_bytes(v); }));
-    }
-    return ::make_shared<value>(buffers, _my_type);
-}
-
-expr::expression lists::delayed_value::to_expression() {
-    std::vector<expr::expression> new_elements;
-    new_elements.reserve(_elements.size());
-
-    for (shared_ptr<term>& e : _elements) {
-        new_elements.emplace_back(expr::to_expression(e));
-    }
-
-    return expr::collection_constructor {
-        .style = expr::collection_constructor::style_type::list,
-        .elements = std::move(new_elements),
-        .type = _my_type
-    };
-}
-
-::shared_ptr<terminal>
-lists::marker::bind(const query_options& options) {
-    const auto& value = options.get_value_at(_bind_index);
-    auto& ltype = dynamic_cast<const list_type_impl&>(_receiver->type->without_reversed());
-    if (value.is_null()) {
-        return nullptr;
-    } else if (value.is_unset_value()) {
-        return constants::UNSET_VALUE;
-    } else {
-        try {
-            value.validate(ltype, options.get_cql_serialization_format());
-            return make_shared<lists::value>(value::from_serialized(value, ltype, options.get_cql_serialization_format()));
-        } catch (marshal_exception& e) {
-            throw exceptions::invalid_request_exception(
-                    format("Exception while binding column {:s}: {:s}", _receiver->name->to_cql_string(), e.what()));
-        }
-    }
-}
-
-expr::expression lists::marker::to_expression() {
-    return expr::bind_variable {
-        .shape = expr::bind_variable::shape_type::scalar,
-        .bind_index = _bind_index,
-        .receiver = _receiver
-    };
-}
-
 void
 lists::setter::execute(mutation& m, const clustering_key_prefix& prefix, const update_parameters& params) {
     auto value = expr::evaluate(*_e, params._options);
