+++ cql3/maps.cc
@@ -41,179 +41,14 @@
 
 #include "maps.hh"
 #include "cql3/abstract_marker.hh"
-#include "cql3/term.hh"
 #include "operation.hh"
 #include "update_parameters.hh"
 #include "exceptions/exceptions.hh"
 #include "cql3/cql3_type.hh"
 #include "constants.hh"
 #include "types/map.hh"
-#include "cql3/tuples.hh"
 
 namespace cql3 {
-
-maps::value
-maps::value::from_serialized(const raw_value_view& fragmented_value, const map_type_impl& type, cql_serialization_format sf) {
-    try {
-        // Collections have this small hack that validate cannot be called on a serialized object,
-        // but compose does the validation (so we're fine).
-        // FIXME: deserialize_for_native_protocol?!
-        auto m = fragmented_value.deserialize<map_type_impl::native_type>(type, sf);
-        std::map<managed_bytes, managed_bytes, serialized_compare> map(type.get_keys_type()->as_less_comparator());
-        if (sf.collection_format_unchanged()) {
-            std::vector<std::pair<managed_bytes, managed_bytes>> tmp = fragmented_value.with_value([sf] (const FragmentedView auto& v) {
-                return partially_deserialize_map(v, sf);
-            });
-            for (auto&& key_value : tmp) {
-                map.insert(std::move(key_value));
-            }
-        } else [[unlikely]] {
-            auto m = fragmented_value.deserialize<map_type_impl::native_type>(type, sf);
-            for (auto&& e : m) {
-                map.emplace(type.get_keys_type()->decompose(e.first),
-                            type.get_values_type()->decompose(e.second));
-            }
-        }
-        return maps::value(std::move(map), type.shared_from_this());
-    } catch (marshal_exception& e) {
-        throw exceptions::invalid_request_exception(e.what());
-    }
-}
-
-cql3::raw_value
-maps::value::get(const query_options& options) {
-    return cql3::raw_value::make_value(get_with_protocol_version(cql_serialization_format::internal()));
-}
-
-managed_bytes
-maps::value::get_with_protocol_version(cql_serialization_format sf) {
-    //FIXME: share code with serialize_partially_deserialized_form
-    size_t len = collection_value_len(sf) * map.size() * 2 + collection_size_len(sf);
-    for (auto&& e : map) {
-        len += e.first.size() + e.second.size();
-    }
-    managed_bytes b(managed_bytes::initialized_later(), len);
-    managed_bytes_mutable_view out(b);
-
-    write_collection_size(out, map.size(), sf);
-    for (auto&& e : map) {
-        write_collection_value(out, sf, e.first);
-        write_collection_value(out, sf, e.second);
-    }
-    return b;
-}
-
-bool
-maps::value::equals(const map_type_impl& mt, const value& v) {
-    return std::equal(map.begin(), map.end(),
-                      v.map.begin(), v.map.end(),
-                      [&mt] (auto&& e1, auto&& e2) {
-        return mt.get_keys_type()->compare(e1.first, e2.first) == 0
-                && mt.get_values_type()->compare(e1.second, e2.second) == 0;
-    });
-}
-
-sstring
-maps::value::to_string() const {
-    // FIXME:
-    abort();
-}
-
-bool
-maps::delayed_value::contains_bind_marker() const {
-    // False since we don't support them in collection
-    return false;
-}
-
-void
-maps::delayed_value::fill_prepare_context(prepare_context& ctx) const {
-}
-
-shared_ptr<terminal>
-maps::delayed_value::bind(const query_options& options) {
-    const map_type_impl& my_map_type = dynamic_cast<const map_type_impl&>(_my_type->without_reversed());
-
-    serialized_compare comparator = my_map_type.get_keys_type()->as_less_comparator();
-    std::map<managed_bytes, managed_bytes, serialized_compare> buffers(std::move(comparator));
-    for (auto&& entry : _elements) {
-        auto&& key = entry.first;
-        auto&& value = entry.second;
-
-        // We don't support values > 64K because the serialization format encode the length as an unsigned short.
-        auto key_bytes = expr::evaluate_to_raw_view(key, options);
-        if (key_bytes.is_null()) {
-            throw exceptions::invalid_request_exception("null is not supported inside collections");
-        }
-        if (key_bytes.is_unset_value()) {
-            throw exceptions::invalid_request_exception("unset value is not supported inside collections");
-        }
-        if (key_bytes.size_bytes() > std::numeric_limits<uint16_t>::max()) {
-            throw exceptions::invalid_request_exception(format("Map key is too long. Map keys are limited to {:d} bytes but {:d} bytes keys provided",
-                                                   std::numeric_limits<uint16_t>::max(),
-                                                   key_bytes.size_bytes()));
-        }
-        auto value_bytes = expr::evaluate_to_raw_view(value, options);
-        if (value_bytes.is_null()) {
-            throw exceptions::invalid_request_exception("null is not supported inside collections");\
-        }
-        if (value_bytes.is_unset_value()) {
-            return constants::UNSET_VALUE;
-        }
-        buffers.emplace(*to_managed_bytes_opt(key_bytes), *to_managed_bytes_opt(value_bytes));
-    }
-    return ::make_shared<value>(std::move(buffers), _my_type);
-}
-
-expr::expression maps::delayed_value::to_expression() {
-    std::vector<expr::expression> new_elements;
-    new_elements.reserve(_elements.size());
-
-    const map_type_impl& mtype = dynamic_cast<const map_type_impl&>(_my_type->without_reversed());
-    data_type ttype = tuple_type_impl::get_instance({mtype.get_keys_type(), mtype.get_values_type()});
-
-    for (auto&& [key, value] : _elements) {
-        expr::expression key_expr = expr::to_expression(key);
-        expr::expression value_expr = expr::to_expression(value);
-        new_elements.emplace_back(expr::tuple_constructor{{std::move(key_expr), std::move(value_expr)}, std::move(ttype)});
-    }
-
-    return expr::collection_constructor {
-        .style = expr::collection_constructor::style_type::map,
-        .elements = std::move(new_elements),
-        .type = _my_type
-    };
-}
-
-::shared_ptr<terminal>
-maps::marker::bind(const query_options& options) {
-    auto val = options.get_value_at(_bind_index);
-    if (val.is_null()) {
-        return nullptr;
-    }
-    if (val.is_unset_value()) {
-        return constants::UNSET_VALUE;
-    }
-    try {
-        val.validate(*_receiver->type, options.get_cql_serialization_format());
-    } catch (marshal_exception& e) {
-        throw exceptions::invalid_request_exception(
-                format("Exception while binding column {:s}: {:s}", _receiver->name->to_cql_string(), e.what()));
-    }
-    return ::make_shared<maps::value>(
-            maps::value::from_serialized(
-                    val,
-                    dynamic_cast<const map_type_impl&>(_receiver->type->without_reversed()),
-                    options.get_cql_serialization_format()));
-}
-
-expr::expression maps::marker::to_expression() {
-    return expr::bind_variable {
-        .shape = expr::bind_variable::shape_type::scalar,
-        .bind_index = _bind_index,
-        .receiver = _receiver
-    };
-}
-
 void
 maps::setter::execute(mutation& m, const clustering_key_prefix& row_key, const update_parameters& params) {
     expr::constant value = expr::evaluate(*_e, params._options);
