+++ cql3/prepare_context.hh
@@ -67,12 +67,10 @@ private:
     std::vector<shared_ptr<column_identifier>> _variable_names;
     std::vector<lw_shared_ptr<column_specification>> _specs;
     std::vector<lw_shared_ptr<column_specification>> _target_columns;
-    // A list of pointers to prepared `function_call` AST nodes, that
+    // A list of pointers to prepared `function_call` cache ids, that
     // participate in partition key ranges computation within an LWT statement.
-    using function_calls_t = std::vector<::shared_ptr<cql3::functions::function_call>>;
-    function_calls_t _pk_fn_calls;
-
     std::vector<::shared_ptr<std::optional<uint8_t>>> _pk_function_calls_cache_ids;
+
     // The flag denoting whether the context is currently in partition key
     // processing mode (inside query restrictions AST nodes). If set to true,
     // then every `function_call` instance will be recorded in the context and
@@ -100,7 +98,6 @@ public:
 
     // Record a new function call, which evaluates a partition key constraint.
     // Also automatically assigns an id to the AST node for caching purposes.
-    void add_pk_function_call(::shared_ptr<cql3::functions::function_call> fn);
     void add_pk_function_call(cql3::expr::function_call& fn);
 
     // Inform the context object that it has started or ended processing the
