+++ cql3/sets.cc
@@ -26,154 +26,6 @@
 #include "types/set.hh"
 
 namespace cql3 {
-
-sets::value
-sets::value::from_serialized(const raw_value_view& val, const set_type_impl& type, cql_serialization_format sf) {
-    try {
-        std::set<managed_bytes, serialized_compare> elements(type.get_elements_type()->as_less_comparator());
-        if (sf.collection_format_unchanged()) {
-            utils::chunked_vector<managed_bytes> tmp = val.with_value([sf] (const FragmentedView auto& v) {
-                return partially_deserialize_listlike(v, sf);
-            });
-            for (auto&& element : tmp) {
-                elements.insert(std::move(element));
-            }
-        } else [[unlikely]] {
-            auto s = val.deserialize<set_type_impl::native_type>(type, sf);
-            for (auto&& element : s) {
-                elements.insert(elements.end(), managed_bytes(type.get_elements_type()->decompose(element)));
-            }
-        }
-        return value(std::move(elements), type.shared_from_this());
-    } catch (marshal_exception& e) {
-        throw exceptions::invalid_request_exception(e.what());
-    }
-}
-
-cql3::raw_value
-sets::value::get(const query_options& options) {
-    return cql3::raw_value::make_value(get_with_protocol_version(cql_serialization_format::internal()));
-}
-
-managed_bytes
-sets::value::get_with_protocol_version(cql_serialization_format sf) {
-    return collection_type_impl::pack_fragmented(_elements.begin(), _elements.end(),
-            _elements.size(), sf);
-}
-
-bool
-sets::value::equals(const set_type_impl& st, const value& v) {
-    if (_elements.size() != v._elements.size()) {
-        return false;
-    }
-    auto&& elements_type = st.get_elements_type();
-    return std::equal(_elements.begin(), _elements.end(),
-            v._elements.begin(),
-            [elements_type] (managed_bytes_view v1, managed_bytes_view v2) {
-                return elements_type->equal(v1, v2);
-            });
-}
-
-sstring
-sets::value::to_string() const {
-    sstring result = "{";
-    bool first = true;
-    for (auto&& e : _elements) {
-        if (!first) {
-            result += ", ";
-        }
-        first = true;
-        result += to_hex(e);
-    }
-    result += "}";
-    return result;
-}
-
-bool
-sets::delayed_value::contains_bind_marker() const {
-    // False since we don't support them in collection
-    return false;
-}
-
-void
-sets::delayed_value::fill_prepare_context(prepare_context& ctx) const {
-}
-
-shared_ptr<terminal>
-sets::delayed_value::bind(const query_options& options) {
-    const set_type_impl& my_set_type = dynamic_cast<const set_type_impl&>(_my_type->without_reversed());
-
-    std::set<managed_bytes, serialized_compare> buffers(my_set_type.get_elements_type()->as_less_comparator());
-    for (auto&& t : _elements) {
-        auto b = expr::evaluate_to_raw_view(t, options);
-
-        if (b.is_null()) {
-            throw exceptions::invalid_request_exception("null is not supported inside collections");
-        }
-        if (b.is_unset_value()) {
-            return constants::UNSET_VALUE;
-        }
-        // We don't support value > 64K because the serialization format encode the length as an unsigned short.
-        if (b.size_bytes() > std::numeric_limits<uint16_t>::max()) {
-            throw exceptions::invalid_request_exception(format("Set value is too long. Set values are limited to {:d} bytes but {:d} bytes value provided",
-                    std::numeric_limits<uint16_t>::max(),
-                    b.size_bytes()));
-        }
-        buffers.insert(buffers.end(), *to_managed_bytes_opt(b));
-    }
-    return ::make_shared<value>(std::move(buffers), _my_type);
-}
-
-expr::expression sets::delayed_value::to_expression() {
-    std::vector<expr::expression> new_elements;
-    new_elements.reserve(_elements.size());
-
-    for (shared_ptr<term>& e : _elements) {
-        new_elements.emplace_back(expr::to_expression(e));
-    }
-
-    return expr::collection_constructor {
-        .style = expr::collection_constructor::style_type::set,
-        .elements = std::move(new_elements),
-        .type = _my_type,
-    };
-}
-
-sets::marker::marker(int32_t bind_index, lw_shared_ptr<column_specification> receiver)
-    : abstract_marker{bind_index, std::move(receiver)} {
-    if (!_receiver->type->without_reversed().is_set()) {
-        throw std::runtime_error(format("Receiver {} for set marker has wrong type: {}",
-                                        _receiver->cf_name, _receiver->type->name()));
-    }
-}
-
-::shared_ptr<terminal>
-sets::marker::bind(const query_options& options) {
-    const auto& value = options.get_value_at(_bind_index);
-    if (value.is_null()) {
-        return nullptr;
-    } else if (value.is_unset_value()) {
-        return constants::UNSET_VALUE;
-    } else {
-        auto& type = dynamic_cast<const set_type_impl&>(_receiver->type->without_reversed());
-        try {
-            value.validate(type, options.get_cql_serialization_format());
-        } catch (marshal_exception& e) {
-            throw exceptions::invalid_request_exception(
-                    format("Exception while binding column {:s}: {:s}", _receiver->name->to_cql_string(), e.what()));
-        }
-        return make_shared<cql3::sets::value>(value::from_serialized(value, type, options.get_cql_serialization_format()));
-    }
-}
-
-expr::expression sets::marker::to_expression() {
-    return expr::bind_variable {
-        .shape = expr::bind_variable::shape_type::scalar,
-        .bind_index = _bind_index,
-        .receiver = _receiver
-    };
-}
-
 void
 sets::setter::execute(mutation& m, const clustering_key_prefix& row_key, const update_parameters& params) {
     expr::constant value = expr::evaluate(*_e, params._options);
