+++ cql3/statements/batch_statement.hh
@@ -42,6 +42,7 @@
 #include "raw/batch_statement.hh"
 #include "timestamp.hh"
 #include "log.hh"
+#include "service_permit.hh"
 
 namespace cql_transport::messages {
     class result_message;
