+++ cql3/user_types.hh
@@ -42,7 +42,6 @@
 
 #include "cql3/abstract_marker.hh"
 #include "column_specification.hh"
-#include "term.hh"
 #include "column_identifier.hh"
 #include "operation.hh"
 #include "to_string.hh"
@@ -56,45 +55,6 @@ class user_types {
     user_types() = delete;
 public:
     static lw_shared_ptr<column_specification> field_spec_of(const column_specification& column, size_t field);
-
-    class value : public terminal {
-        std::vector<managed_bytes_opt> _elements;
-    public:
-        explicit value(std::vector<managed_bytes_opt>, data_type my_type);
-
-        static value from_serialized(const raw_value_view&, const user_type_impl&);
-
-        virtual cql3::raw_value get(const query_options&) override;
-        virtual sstring to_string() const override;
-    };
-
-    // Same purpose than Lists.DelayedValue, except we do handle bind marker in that case
-    class delayed_value : public non_terminal {
-        user_type _type;
-        std::vector<shared_ptr<term>> _values;
-    public:
-        delayed_value(user_type type, std::vector<shared_ptr<term>> values);
-        virtual bool contains_bind_marker() const override;
-        virtual void fill_prepare_context(prepare_context& ctx) const override;
-    private:
-        std::vector<managed_bytes_opt> bind_internal(const query_options& options);
-    public:
-        virtual shared_ptr<terminal> bind(const query_options& options) override;
-
-        virtual expr::expression to_expression() override;
-    };
-
-    class marker : public abstract_marker {
-    public:
-        marker(int32_t bind_index, lw_shared_ptr<column_specification> receiver)
-            : abstract_marker{bind_index, std::move(receiver)}
-        {
-            assert(_receiver->type->is_user_type());
-        }
-
-        virtual shared_ptr<terminal> bind(const query_options& options) override;
-        virtual expr::expression to_expression() override;
-    };
 
     class setter : public operation {
     public:
